+++ crates/dslab-faas/src/extra/mod.rs
@@ -4,4 +4,3 @@ pub mod hermes;
 pub mod hybrid_histogram;
 pub mod opendc_trace;
 pub mod resolvers;
-pub mod simple_schedulers;
