+++ examples/faas-parallel/readme.md
@@ -1,2 +1,2 @@
 # Parallel FaaS example
-This crate shows how to setup several parallel simulations using ParallelHostConfig. An alternative way is to use YAML config, as it is done in Serverless in the Wild reproduction experiment.
+This example demonstrates how to setup several parallel DSLab FaaS simulations using `ParallelHostConfig`. An alternative way is to use YAML config, as it is done in the [Serverless in the Wild](../serverless-in-the-wild) reproduction experiment.
