+++ examples/faas-scheduling-experiment/src/main.rs
@@ -12,7 +12,7 @@ use dslab_faas::stats::Stats;
 
 #[derive(Serialize, Deserialize)]
 struct ExperimentConfig {
-    pub base_config: RawConfig,
+    pub sim_config: RawConfig,
     pub schedulers: Vec<String>,
 }
 
@@ -51,7 +51,7 @@ fn main() {
     let experiment_config: ExperimentConfig =
         serde_yaml::from_reader(File::open(Path::new(&args[2])).unwrap()).unwrap();
     let schedulers = experiment_config.schedulers;
-    let sim_config = experiment_config.base_config;
+    let sim_config = experiment_config.sim_config;
     let configs: Vec<_> = schedulers
         .iter()
         .map(|x| {
