+++ examples/serverless-in-the-wild/config.yaml
@@ -1,4 +1,4 @@
-base_config:
+sim_config:
   hosts:
     - cores: 8
       resources:
